requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
websocket-client==1.6.4
eventlet==0.33.3
requests==2.31.0
gunicorn==21.2.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.2.1
redis==5.0.1
flask-cors==4.0.0
setuptools>=68



